Break most of insights page into two separate components
